refactor(tools): extract image domain into src/tools/image.py
